Add "line" input format
This makes it easier to use klp with any text file
(which could be done before with "-k _line"), for example
to filter by Python expression (--where) or transform the output
(--output-eval).

Uses "line" as the key, because I find that quite intuitive.
But a case could be made for "message" or "msg", because that
would produce output with with -c.

Keeps leading whitespace, because that could have semantic meaning
(e.g. in YAML or Python or Haskell).