Use _klp_* prefix for synthetic fields
Helps avoid collisions with existing keys,
makes it clearer that this a field generated by klp.

Use _klp_timedelta instead of _klp_ts_delta, because this
is the delta between the original timestamps in the file,
not between the synthetic _klp_ts values.

Keep _ts as a TS_KEY because it might be used in real logs.