Add split_startswith()/split_endswith() functions
These should be useful for "-f data" to split the whole data
into events, where a naive splitlines() is not sufficient.

Note that we return text before the first match and after the last
match, so that no info is suppressed. This might not be clear
from the function names, but I guess this is the most useful
behaviour in practice. We just want an intelligent split,
not a filter.