Add keep_all options to split_endswith/startswith
Setting this to False returns only those parts that really end/start
with the given pattern.

But usually we return the part before the first startswith and
the part after the last endswith, too, in order to act like split() and
not lose any data. So this is still the default.